Commit #2. Add proof
